pyproject.toml
[project]
name = "sre-bot-strands"
version = "0.1.0"
description = "AWS Strands FastAPI with Two-Service Architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "strands-agents==1.10.0",
    "strands-agents-tools==0.2.9",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.0.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
]

# Note: slack-bolt can be added later when implementing Slack integration

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx>=0.27.0",  # for FastAPI TestClient
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
